# backend-marketing/openapi.yaml
openapi: 3.0.3
info:
  title: FeedPlug API
  description: API Backend for FeedPlug Marketing Platform
  version: 1.0.0
  contact:
    name: FeedPlug Support
servers:
  - url: /api/v1
    description: Current API version
...
paths:
...
  /accounts:
    get:
      tags: [Accounts]
      summary: Get account info
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '403':
          description: No account associated

    put:
      tags: [Accounts]
      summary: Update account
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Account updated
  /accounts/users:
    get:
      tags: [Accounts]
      summary: List account users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /accounts/users/invite:
    post:
      tags: [Accounts]
      summary: Invite user to account
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                firstName:
                  type: string
                lastName:
                  type: string
                role:
                  type: string
                  enum: [OWNER, MANAGER, VIEWER, AGENCY]
      responses:
        '201':
          description: Invitation sent
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        accountId:
          type: string
        accountName:
          type: string
        plan:
          type: string
    Account:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        plan:
          type: string
        email:
          type: string
        trialEndsAt:
          type: string
          format: date-time
        billingStatus:
          type: string
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
...